separated `ImagePick` widget into a file.
